dagster-pipelines/pyproject.toml
[project]
name = "dagster-pipelines"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "dagster",
    "dagster-webserver",
    "dagster-postgres",
    "pandas==2.2.3",
    "pyarrow==18.1.0",
    "lightgbm==4.5.0",
    "scikit-learn==1.6.1",
    "mlflow==2.19.0",
    "boto3==1.42.59",
    "psycopg2-binary==2.9.11",
    "feast",
    "evidently",
    "requests",
    "deltalake",
    "pandera",
    "duckdb",
    "pyyaml",
]
